Multiagent NFL/requirements.txt
langgraph>=0.2.30
langchain>=0.2.15
langchain-openai>=0.1.22
langchain-community>=0.2.12
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
tiktoken>=0.7.0